+++ java/src/main/java/com/google/api/services/samples/authorizedbuyers/realtimebidding/v1/bidders/publisherConnections/BatchApprovePublisherConnections.java
@@ -39,7 +39,7 @@ import net.sourceforge.argparse4j.inf.Namespace;
 public class BatchApprovePublisherConnections {
 
   public static void execute(RealTimeBidding client, Namespace parsedArgs) throws IOException {
-    Integer bidderAccountId = parsedArgs.getInt("account_id");
+    Long bidderAccountId = parsedArgs.getLong("account_id");
     String parent = String.format("bidders/%d", bidderAccountId);
 
     String publisherConnectionNameTemplate = "bidders/%d/publisherConnections/%s";
@@ -80,7 +80,7 @@ public class BatchApprovePublisherConnections {
             "The resource ID of the bidders resource for which the publisher connections are "
                 + "being approved.")
         .required(true)
-        .type(Integer.class);
+        .type(Long.class);
     parser
         .addArgument("-p", "--publisher_connection_ids")
         .help(
